backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
langdetect==1.0.9
pydantic-settings==2.7.1
python-dotenv==1.0.1
